backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pdfplumber
python-docx
python-multipart
